Fragmentation access rights

+++ app/controllers/projects/point_files_controller.rb
@@ -1,5 +1,6 @@
 class Projects::PointFilesController < ApplicationController
   before_action :authenticate_user!
+  after_action :verify_authorized, except: [:datatables_index_zs_point, :datatables_index_ww_point]
   before_action :set_point_file, only: [:show, :edit, :update, :destroy]
 
 
@@ -20,7 +21,7 @@ class Projects::PointFilesController < ApplicationController
 
   def download
     @point_file = PointFile.find(params[:id])
-#    attachment_authorize(:attachment, "show", @attachment.attachmenable_type.singularize.downcase)    
+    authorize @point_file, :download?
     send_file "#{@point_file.load_file.file.file}", 
       type: "#{@point_file.load_file.file.content_type}",
       x_sendfile: true
@@ -30,6 +31,7 @@ class Projects::PointFilesController < ApplicationController
   # GET /point_files/1.json
   def show
     @project = load_project
+    authorize @point_file, :show?
   end
 
   # GET /point_files/new
@@ -38,12 +40,13 @@ class Projects::PointFilesController < ApplicationController
     @point_file = PointFile.new
     @point_file.project = @project
     @point_file.load_date = Time.zone.today 
-#    authorize :point_file, :new?
+    authorize :point_file, :new?
   end
 
   # GET /point_files/1/edit
   def edit
     @project = load_project
+    authorize @point_file, :edit?
   end
 
   # POST /point_files
@@ -52,7 +55,7 @@ class Projects::PointFilesController < ApplicationController
     @project = load_project
     @point_file = PointFile.new(point_file_params)
     @point_file.project = @project
-#    authorize @point_file, :create?
+    authorize @point_file, :create?
     respond_to do |format|
       if @point_file.save
         flash[:success] = t('activerecord.successfull.messages.created', data: @point_file.fullname)
@@ -71,7 +74,7 @@ class Projects::PointFilesController < ApplicationController
   # PATCH/PUT /point_files/1.json
   def update
     @project = load_project
-#    authorize @point_file, :update?
+    authorize @point_file, :update?
     respond_to do |format|
       if @point_file.update(point_file_params)
         flash[:success] = t('activerecord.successfull.messages.updated', data: @point_file.fullname)
@@ -88,7 +91,7 @@ class Projects::PointFilesController < ApplicationController
   # DELETE /point_files/1.json
   def destroy
     @project = load_project
-#    authorize @point_file, :destroy?
+    authorize @point_file, :destroy?
     if @point_file.destroy
       flash[:success] = t('activerecord.successfull.messages.destroyed', data: @point_file.fullname)
       redirect_to project_path(@project)
